requirements.txt
telethon>=1.11.3,<2.0.0
psycopg2>=2.8.4,<3.0.0
cryptg>=0.2,<1.0
python-dotenv>=0.12.0,<1.0.0
aiohttp>=3.6.2,<4.0.0
